+++ apps/elf-council-frontend/src/ui/proposals/ProposalsDetailsCard/GSCMember.tsx
