+++ apps/elf-council-frontend/src/ui/proposals/ProposalsDetailsCard/GSCVoteTally.tsx
