+++ apps/elf-council-frontend/src/ui/proposals/ProposalsDetailsCard/StaleVotingPowerMessage.tsx
