+++ apps/elf-council-frontend/src/ui/proposals/ProposalsDetailsCard/UnverifiedProposalWarning.tsx
