+++ apps/elf-council-frontend/src/ui/proposals/ProposalsDetailsCard/tooltipDefinitions.ts
